modified:   package-lock.json 	modified:   package.json 	new file:   src/components/registration/EventRegCard.tsx 	modified:   src/components/registration/test.tsx
