+++ package.json
@@ -38,6 +38,7 @@
     "next": "14.2.5",
     "next-themes": "^0.3.0",
     "nodemailer": "^6.9.14",
+    "qrcode": "^1.5.4",
     "quill": "^2.0.2",
     "react": "^18",
     "react-dom": "^18",
@@ -60,6 +61,7 @@
     "@types/formidable": "^3.4.5",
     "@types/node": "^20.16.1",
     "@types/nodemailer": "^6.4.15",
+    "@types/qrcode": "^1.5.5",
     "@types/react": "^18.3.4",
     "@types/react-dom": "^18",
     "@types/uuid": "^10.0.0",
